requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorclass==2.2.2
distlib==0.3.6
Django==4.2.2
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
doctor==3.13.7
filelock==3.12.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
jsonschema==2.6.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.1
pbr==5.11.1
Pillow==9.5.0
pip-upgrader==1.4.15
pipreqs==0.4.13
platformdirs==3.8.0
PyJWT==2.3.0
python-dotenv==1.0.0
pytz==2021.3
PyYAML==3.13
requests==2.31.0
rfc3987==1.3.8
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.2
stevedore==5.1.0
strict-rfc3339==0.7
terminaltables==3.1.10
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2021.5
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
yarg==0.1.9

# pip install -r requirements.txt
# pip freeze > requirements.txt